+++ CHANGELOG.md
@@ -9,13 +9,26 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Font size slider
+
+### Changed
+
+- Rearranged settings view
+- Decreased default font size
+- Using github styled markdown
+
+
 ## [2023.11.22]
 
 ### Fixed
+
 - Stutter caused by phase change actions
 - Stutter cause by @Environment Dismiss
 
 ### Misc
+
 - Code clean up
 - Removed unused code and comments
 - Added SwiftLint and Periphery run script and schemes